Linked list sort tests now also use pthreads. Time measurements and printing are done using unsigned long longs to avoid overflow.

+++ src/utils.h
@@ -3,7 +3,7 @@
 #include <sys/time.h>
 
 int get_random_num(const int a, const int b);
-long get_time_diff(struct timespec start, struct timespec stop);
+unsigned long long get_time_diff(struct timespec start, struct timespec stop);
 void wait_for_enter();
 
 #endif
